Delete non-existent images from dest on image sync (to staging)
We periodically run the sync-images script (as a heroku scheduler
job) that syncs all image changes from production to staging.
However, this script only performs PUT and POST. There is no
DELETE.

The consequence is that we have lots of old images on staging
that no longer exist in production or on GCE. They have already
been deleted by gcloud-cleanup. That then causes builds to no
longer start (on staging), and cycle between created/queued
states.

This patch makes sure we perform a DELETE for those images, to
truly keep staging in sync with production.